requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-allauth==65.3.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-decouple==3.8
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
cryptography==46.0.2
